% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlm.R
\name{ggplot_lm}
\alias{ggplot_lm}
\title{Plot linear regression model}
\usage{
ggplot_lm(mod)
}
\arguments{
\item{mod}{linear regression model.}
}
\description{
Plot linear regression model with confidence and prediction intervals
}
