% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlm.R
\name{make_intervals}
\alias{make_intervals}
\title{Creates new data and prediction/confidence intervals}
\usage{
make_intervals(mod, steps = 100)
}
\arguments{
\item{mod}{linear regression model}

\item{steps}{how many new data points}
}
\description{
Creates new data and prediction/confidence intervals
}
